+++ docs/fr_FR/Changelog.rst
@@ -3,7 +3,7 @@ ChangeLog
 
 - Interne: Msg 8012 n'est plus transmis à 'AbeilleCmd'.
 - Interne: Cmd: Correction changement adresse lors d'une nouvelle annonce.
-- Interne: Correction status 'NO-ACK'.
+- Interne: Correction status 'NO-ACK' + 'ab::noack' => 'ab::txAck'.
 
 231009-BETA-1
 -------------